requirements.txt
picamera
RPi.GPIO
cffi
numpy
imutils
